Change shopping list item insertion to be case and whitespace insensitive. Add clear all items in shopping list and add confirmation prompts to all delete operations

+++ src/app/recipes/recipe-detail/recipe-detail.component.ts
@@ -84,6 +84,14 @@ export class RecipeDetailComponent implements OnInit {
   }
 
   onDeleteRecipe() {
+    const dialogAccepted = confirm(
+      'Are you sure you want to do this? The recipe will be deleted permanently.'
+    );
+
+    if (!dialogAccepted) {
+      return;
+    }
+
     this.isProcessingAction = true;
     this.dataStorageService.deleteRecipe(this.recipe.id).subscribe(() => {
       this.recipeService.deleteRecipe(this.id);
